Implemented the conversion from skn_file to a static md2_file (mesh, texture, light normals) and rspecced it with model Akali, Annie and her bear.
Also with:
1) small modification to skl_file
2) Improved md2_file_spec, added get_anorm_index for md2_file

NOTE:

Codes are just working, but ugly, inefficient and possibly buggy. The md2 files generated are tested with GLGE's md2_demo, but nothing else.

Due to copy right issues, the used and generated models are not comming with this commit.

All specs passed in 368.86 seconds. That's mostly used on orgo.md2.